requirements.txt
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
pyjwt==2.8.0
cryptography==41.0.7
passlib==1.7.4
python-can==4.3.1
securesystemslib==0.28.0
tuf==2.0.0
bcrypt==4.1.1
uvicorn==0.27.1
fastapi==0.109.2
